Feature/data base added

+++ Encounter Me/TrailContainer.cs
@@ -5,7 +5,7 @@ public class TrailContainer
 	public int Id { get; set; }
 	public double Lat { get; set; }
 	public double Lng { get; set; }
-	public float Length { get; set; }
+	public double Length { get; set; }
 	public int Diff { get; set; }
 	public string GeoJsonData { get; set; }
 	public TrailType trailType { get; set; }
